A bunch of cleanups... (#43)
* A bunch of cleanups...

- module update
- more/fixed docs
- better docker integration

* fix jenkins trigger with "\" in names: "trigger job prefix/jobName"

+++ bot/stats/stats_test.go
@@ -8,15 +8,15 @@ import (
 func TestStats(t *testing.T) {
 	value, err := Get("test")
 	assert.NotNil(t, err)
-	assert.Equal(t, value, 0)
+	assert.Equal(t, value, uint(0))
 
 	Increase("test", 2)
 	value, err = Get("test")
 	assert.Nil(t, err)
-	assert.Equal(t, value, 2)
+	assert.Equal(t, value, uint(2))
 
 	IncreaseOne("test")
 	value, err = Get("test")
 	assert.Nil(t, err)
-	assert.Equal(t, value, 3)
+	assert.Equal(t, value, uint(3))
 }
